\name{dat1p}
\alias{dat1p}
\docType{data}

\title{
A sample data oobject conatining a list of parameters for function httk.fun
}
\description{
An example for list of parameters httk.fun awaits.}

\usage{data("dat1p")}

\format{
A lsit of four objects 
 $ chem.name: chr "imazalil"
 $ species  : chr "human"
 $ days     : int 10
 $ dose     : NULL
}

\details{
A list of parameters for httk.fun}

\source{
A toy example for httk.fun based on jaqpot specifications
}
\references{
httk cran package
}
\examples{
data(dat1p)
## maybe str(dat1p) ; plot(dat1p) ...
}
\keyword{datasetsp}
